разбил открытие и валидацию
