+++ js/data.js
@@ -1,20 +1,20 @@
 import {getRandomInteger, createRandomIdFromRangeGenerator} from "./util.js";
 
 const DESCRIPTION = [
-  "Лучше быть последним — первым, чем первым — последним.",
-  "Каждый в цирке думает, что знает в цирке, но не каждый, что в цирке знает, что в цирке не каждый знает думает.",
-  "За двумя зайцами погонишься — рыбку из пруда не выловишь, делу время, а отмеришь семь раз…",
-  "Кем бы ты ни был, кем бы ты не стал, помни, где ты был и кем ты стал.",
-  "Иногда жизнь — это жизнь, а ты в ней иногда."
+  'Лучше быть последним — первым, чем первым — последним.',
+  'Каждый в цирке думает, что знает в цирке, но не каждый, что в цирке знает, что в цирке не каждый знает думает.',
+  'За двумя зайцами погонишься — рыбку из пруда не выловишь, делу время, а отмеришь семь раз…',
+  'Кем бы ты ни был, кем бы ты не стал, помни, где ты был и кем ты стал.',
+  'Иногда жизнь — это жизнь, а ты в ней иногда.'
 ]
 
 const MESSAGE = [
-  "Всё отлично!",
-  "В целом всё неплохо. Но не всё.",
-  "Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.",
-  "Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.",
-  "Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.",
-  "Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!"
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
 ]
 
 const NAMES = [
